tests,dispatch: Move some functions to src/nvim/lib/helpers.lua
